app/schemas/project.py
from pydantic import BaseModel

class ProjectCreate(BaseModel):
    name: str
    description: str | None = None

class ProjectResponse(BaseModel):
    id: int
    name: str
    description: str | None

    class Config:
        from_attributes = True
        
class AddMemberRequest(BaseModel):
    user_id: int

