requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-jose
email-validator
passlib==1.7.4
bcrypt==4.0.1
